requirements.txt
tensorflow==2.13.0
numpy==1.24.3
Pillow==9.5.0
streamlit==1.24.0
matplotlib>=3.4.0
